+++ vpmobil/rework.py
@@ -1,21 +1,31 @@
 from __future__ import annotations
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from pathlib import Path
 from xml.etree import ElementTree as XML
 from typing import Literal
-from datetime import datetime, date
+from datetime import datetime, date, time
 import re
 
 from vpmobil.utils import prettyxml
 
-@dataclass
-class VertretungsTagBase():
+@dataclass(init=True)
+class VpmobilPyModell():
+
+    _data:    XML.Element            = field(init=True)
+    _planart: Literal["K", "L", "R"] = field(init=True)
+
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                   VertretungsTagBase                                     │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class VertretungsTagBase(VpmobilPyModell):
     """Base-Class für Vertretungspläne an einem bestimmten Tag.
 
     Beim Versuch einer Instanzierung wird automatisch eine Instanz von `VertretungsTag`, `LehrerVertretungsTag` oder `RaumVertretungsTag` zurückgegeben.
     """
 
-    _data: XML.ElementTree
+    _data:    XML.ElementTree        = field(init=True)
+    _planart: Literal["K", "L", "R"] = field(init=False, default=None)
 
     def __new__(cls, _data: XML.ElementTree):
         if _data.find(".//planart") is None or _data.find(".//planart").text is None:
@@ -31,13 +41,11 @@ class VertretungsTagBase():
             case _:
                 raise ValueError
             
+    def __post_init__(self):
+        self._planart = self._data.find("Kopf/planart").text
+            
     def __repr__(self):
-        return f"<Vertretungsplan (Typ {self.planart}) vom {self.datum.strftime('%d.%m.%Y')}>"
-    
-    @property
-    def planart(self) -> Literal['K', 'L', 'R']:
-        "Interne Bezeichnung für die Art des Inhalts des Plans"
-        return self._data.find(".//planart").text
+        return f"<Vertretungsplan (Typ {self._planart}) vom {self.datum.strftime('%d.%m.%Y')}>"
         
     @property
     def zeitstempel(self) -> datetime | None:
@@ -62,6 +70,20 @@ class VertretungsTagBase():
             year, month, day = map(int, match.groups())
             return date(year, month, day)
         return None
+    
+    @property
+    def freieTage(self) -> list[date] | None:
+        "Im Vertretungsplan als frei markierte Tage"
+
+        freieTage = self._data.find("FreieTage")
+        if freieTage is None:
+            return None
+        
+        freieTageList: list[date] = []
+        for ft in freieTage.findall("ft"):
+            if ft.text is not None:
+                freieTageList.append(datetime.strptime(ft.text, "%y%m%d").date())
+        return freieTageList
             
     @classmethod
     def fromfile(cls, pfad: Path) -> VertretungsTag | LehrerVertretungsTag | RaumVertretungsTag:
@@ -108,14 +130,398 @@ class VertretungsTagBase():
 
         zielpfad.write_text(xmlpretty, encoding="utf-8")
 
+    def _elemente_Klassen(self) -> list[XML.Element] | None:
+        klassen: list[XML.Element] = []
+        klassen_elemente = self._data.findall('.//Kl')
+        if klassen_elemente is not []:
+            for kl in klassen_elemente:
+                if kl.find('Kurz') is not None:
+                    klassen.append(kl)
+            return klassen
+        return None
+
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                      VertretungsTag                                      │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
 
 class VertretungsTag(VertretungsTagBase):
-    ...
+
+    @property
+    def lehrerKrank(self) -> list[str]:
+        "Aller Lehrer, die unplanmäßig keinen Unterricht haben"
+        
+        lehrerMitUnterricht: set[str] = set()
+        lehrerVielleichtKrank: set[str] = set()
+
+        for klasse in self.klassen:
+            for stunde in [stunde for stunden in klasse.stundenHeute.values() for stunde in stunden]:
+
+                if stunde.ausfall and klasse.kurs(stunde.kursnummer) is not None:
+                    lehrerVielleichtKrank.add(klasse.kurs(stunde.kursnummer).lehrer)
+
+                elif stunde.lehrergeändert:
+                    if stunde.lehrer is not None:
+                        lehrerMitUnterricht.update(stunde.lehrer)
+                    if klasse.kurs(stunde.kursnummer) is not None:
+                        lehrerVielleichtKrank.add(klasse.kurs(stunde.kursnummer).lehrer)
+
+                elif not stunde.ausfall and not stunde.lehrergeändert:
+                    if stunde.lehrer is not None:
+                        lehrerMitUnterricht.update(stunde.lehrer)
+
+        return sorted(
+            {
+                lehrer for lehrer in lehrerVielleichtKrank
+                if lehrer not in lehrerMitUnterricht
+                and lehrer != ""
+                and lehrer is not None
+            }
+        )
+    
+    @property
+    def klassen(self) -> list[Klasse]:
+        "Im Vertretungsplan hinterlegte Klassen"
+        return [Klasse(element, self._planart) for element in (self._elemente_Klassen() or [])]
+    
+    def klasse(self, kürzel: str) -> Klasse | None:
+        "Gibt die Klasse zurück, deren Tag `<Kurz>` gleich `kürzel` ist"
+
+        for kl in self.klassen:
+            if kl.kürzel == kürzel:
+                return kl
+        return None
+
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                   LehrerVertretungsTag                                   │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
 
 class LehrerVertretungsTag(VertretungsTagBase):
-    ...
+
+    @property
+    def lehrer(self) -> list[Lehrer]:
+        "Im Vertretungsplan hinterlegte Klassen"
+        return [Lehrer(element, self._planart) for element in (self._elemente_Klassen() or [])]
+    
+    def get_lehrer(self, kürzel: str) -> Lehrer | None:
+        "Gibt den Lehrer zurück, dessen Tag `<Kurz>` gleich `kürzel` ist"
+
+        for le in self.lehrer:
+            if le.kürzel == kürzel:
+                return le
+        return None
+    
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                    RaumVertretungsTag                                    │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
 
 class RaumVertretungsTag(VertretungsTagBase):
+    
+    @property
+    def räume(self) -> list[Raum]:
+        "Im Vertretungsplan hinterlegte Klassen"
+        return [Raum(element, self._planart) for element in (self._elemente_Klassen() or [])]
+    
+    def raum(self, kürzel: str) -> Raum | None:
+        "Gibt den Lehrer zurück, dessen Tag `<Kurz>` gleich `kürzel` ist"
+
+        for ra in self.räume:
+            if ra.kürzel == kürzel:
+                return ra
+        return None
+    
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                      KlasseLikeBase                                      │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class KlasseLikeBase(VpmobilPyModell):
+    
+    @property
+    def kürzel(self) -> str:
+        return self._data.find('Kurz').text
+    
+    @property
+    def stundenHeute(self) -> dict[int, list[Stunde]]:
+        """Alle Stunden an dem Tag als Dictionary<br>
+        Die Schlüssel sind die Unterrichtsperioden, die Werte Listen von Unterrichsstunden
+        """
+
+        fin: dict[int, list[Stunde]] = {}
+        pl = self._data.find("Pl")
+        for std in pl.findall("Std"):
+            stunde = Stunde(std, self._planart, self.kürzel)
+            nr = stunde.periode
+            if nr is not None:
+                if fin.get(stunde.periode) is None:
+                    fin[stunde.periode] = [stunde]
+                else:
+                    fin[stunde.periode].append(stunde)
+        return fin
+
+    def stundenHeuteInPeriode(self, periode: int) -> list[Stunde]:
+        "Gibt die Stunden an dem Tag in einer bestimmten Unterrichtsperiode zurück"
+        return self.stundenHeute.get(periode) or []
+    
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                           Klasse                                         │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class Klasse(KlasseLikeBase):
+    """Klasse, die den Vertretungsplan für eine bestimmte Klasse an einem bestimmten Tag repräsentiert.
+    
+    Unterstützt Subskription: 
+    ```
+    data: Klasse = vpday.klasse("10a")
+    stunden_zur_dritten = data[3]
+    ```
+    """
+
+    @property
+    def kurse(self) -> list[Kurs]:
+        "Alle im Plan vermerkten Kurse, die die Klasse hat"
+        fin: list[Kurs] = []
+        unterricht = self._data.find("Unterricht")
+        for ue in unterricht.findall("Ue"):
+            fin.append(Kurs(ue.find("UeNr"), self._planart))
+        return fin
+    
+    def kurs(self, kursnummer: int) -> Kurs | None:
+        "Gibt den Kurs der Klasse mit der Kursnummer `kursnummer` zurück"
+        for kurs in self.kurse:
+            if kurs.kursnummer == kursnummer:
+                return kurs
+        return None
+    
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                           Lehrer                                         │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+    
+class Lehrer(KlasseLikeBase):
+    
+    @property
+    def aufsichten(self) -> list[Aufsicht]:
+        """Alle Aufsichten an dem Tag als Dictionary<br>
+        Die Schlüssel sind die Unterrichtsperioden, die Werte Listen von Unterrichsstunden
+        """
+
+        fin: list[Stunde] = []
+        aufsichten = self._data.find("Aufsichten")
+        for aufsicht in aufsichten.findall("Aufsicht"):
+            fin.append(Aufsicht(aufsicht, self._planart))
+        return fin
+
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                           Raum                                           │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class Raum(KlasseLikeBase):
     ...
 
-test = VertretungsTagBase()
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                         Aufsicht                                         │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class Aufsicht(VpmobilPyModell):
+
+    @property
+    def vorStunde(self) -> int:
+        return self._data.find("AuVorStunde").text
+    
+    @property
+    def uhrzeit(self) -> time:
+        "Uhrzeit der Aufsicht"
+        return datetime.strptime(self._data.find("AuUhrzeit").text, "%H:%M").time()
+    
+    @property
+    def zeit(self) -> str:
+        "Hinweis zum Zeitpunkt der Aufsicht"
+        return self._data.find("AuZeit").text
+    
+    @property
+    def ort(self) -> str:
+        "Hinweis zum Ort der Aufsicht"
+        return self._data.find("AuOrt").text
+
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                          Stunde                                          │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class Stunde(VpmobilPyModell):
+
+    _quelle: str = field(init=True)
+    
+    def __repr__(self):
+        if self.ausfall:
+            return f"<Ausfall: '{self.info}'>"
+        return f"<'{self.fach}' bei '{", ".join(self.lehrer)}' in '{", ".join(self.räume)}'>"
+    
+    @property
+    def periode(self) -> int:
+        "Unterrichtsperiode der Stunde"
+        return int(self._data.find("St").text)
+
+    @property
+    def beginn(self) -> time:
+        "Beginn der Stunde"
+        return datetime.strptime(self._data.find("Beginn").text, "%H:%M").time()
+    
+    @property
+    def ende(self) -> time:
+        "Ende der Stunde"
+        return datetime.strptime(self._data.find("Ende").text, "%H:%M").time()
+    
+    @property
+    def ausfall(self) -> bool:
+        "Ob die Stunde entfällt"
+        return self._data.find("Fa").text == "---"
+
+    @property
+    def fach(self) -> str | None:
+        """Fach der Stunde<br>
+        Gibt `None` zurück, wenn die Stunde entfällt
+
+        Es kann sein, dass nicht das wirkliche Fach sondern die Kursbezeichnung zurückgegeben wird. Stattdessen `klasse.kurs(stunde.kursnummer).fach` verwenden.<br>
+        Bei Unsicherheit mit Fallback:
+        ```
+        stunde.fach if klasse.kurs(stunde.kursnummer) is None else klasse.kurs(stunde.kursnummer).fach
+        ```
+        """
+        if self._data.find("Fa") is not None and self._data.find("Fa").text not in [None, "---"]:
+            return self._data.find("Fa").text
+        else:
+            return None
+
+    @property
+    def klassen(self) -> str | None:
+        """Alle Klassen der Stunde<br>
+        Gibt `[]` zurück, wenn die Stunde entfällt oder keine Klassen eingetragen sind
+        """
+        if self._planart == "K":
+            return self._quelle
+        elif self._planart == "R":
+            if self._data.find("Ra") is not None and self._data.find("Ra").text is not None:
+                return self._data.find("Ra").text.split(" ")
+            else:
+                return []
+        elif self._planart == "L":
+            if self._data.find("Le") is not None and self._data.find("Le").text is not None:
+                return self._data.find("Le").text.split(" ")
+            else:
+                return []
+        
+    @property
+    def lehrer(self) -> list[str]:
+        """Alle Lehrer der Stunde<br>
+        Gibt `[]` zurück, wenn die Stunde entfällt oder keine Lehrer eingetragen sind
+        """
+        if self._planart == "L":
+            return self._quelle
+        else:
+            if self._data.find("Le") is not None and self._data.find("Le").text is not None:
+                return self._data.find("Le").text.split(" ")
+            else:
+                return []
+
+        
+    @property
+    def räume(self) -> list[str]:
+        """Räume der Stunde<br>
+        Gibt `[]` zurück, wenn die Stunde entfällt oder keine Räume eingetragen sind
+        """
+        if self._planart == "R":
+            return self._quelle
+        else:
+            if self._data.find("Ra") is not None and self._data.find("Ra").text is not None:
+                return self._data.find("Ra").text.split(" ")
+            else:
+                return []
+        
+    @property
+    def fachgeändert(self) -> bool:
+        "Ob eine Änderung des Fachs für die Stunde vorliegt<br>Ebenfalls `True`, wenn die Stunde entfällt"
+        return "FaAe" in self._data.find("Fa").attrib
+    
+    @property
+    def lehrergeändert(self) -> bool:
+        "Ob eine Änderung des Lehrers für die Stunde vorliegt<br>Ebenfalls `True`, wenn die Stunde entfällt"
+        return "LeAe" in self._data.find("Le").attrib if self._planart != "L" else False
+    
+    @property
+    def raumgeändert(self) -> bool:
+        "Ob eine Änderung des Raums für die Stunde vorliegt<br>Ebenfalls `True`, wenn die Stunde entfällt"
+        return "RaAe" in self._data.find("Ra").attrib if self._planart != "R" else False
+    
+    @property
+    def klassegeändert(self) -> bool:
+        "Ob eine Änderung der Klasse für die Stunde vorliegt<br>Ebenfalls `True`, wenn die Stunde entfällt"
+        if self._planart == "K":
+            return False
+        elif self._planart == "L":
+            return "LeAe" in self._data.find("Le").attrib
+        elif self._planart == "R":
+            return "RaAe" in self._data.find("Ra").attrib
+
+    @property
+    def geändert(self) -> bool:
+        "Ob eine Änderung im Plan vorliegt<br>Ebenfalls `True`, wenn die Stunde entfällt"
+        return self.fachgeändert or self.lehrergeändert or self.raumgeändert or self.klassegeändert
+
+    @property
+    def kursnummer(self) -> int | None:
+        """Nummer des Kurses der Stunde<br>
+        Kann `None` sein, wenn das Fach der Stunde geändert wurde, jedoch nicht, wenn die Stunde entfällt.<br>
+        Kann `None` sein, beispielsweise wenn die Stunde eine Exkursion ist.
+        
+        Kursnummern können verwendet werden, um in den Kursen einer Klasse mehr Details zu einem Kurs zu erhalten, beispielsweise, wenn eine Unterrichtsstunde ausfällt und Informationen wie Lehrer, Fach und Raum deswegen nicht verfügbar sind.
+        """
+        if self._data.find("Nr") is not None and self._data.find("Nr").text is not None:
+            return int(self._data.find("Nr").text)
+        else:
+            return None
+    
+    @property
+    def info(self) -> str | None:
+        "Zusätzliche Information der Stunde"
+        if self._data.find("If") is not None and self._data.find("If").text is not None and self._data.find("If").text != "":
+            return self._data.find("If").text
+        else:
+            return None
+        
+# ╭──────────────────────────────────────────────────────────────────────────────────────────╮
+# │                                          Kurs                                            │ 
+# ╰──────────────────────────────────────────────────────────────────────────────────────────╯
+
+class Kurs(VpmobilPyModell):
+    """Klasse die einen bestimmten Kurs repräsentiert.
+    """
+
+    def __repr__(self) -> str:
+        return f"<'{self.fach}' bei '{self.lehrer}', Gruppe '{self.gruppe or '-'}' (Kursnummer '{self.kursnummer}')>"
+    
+    @property
+    def lehrer(self) -> str | None:
+        "Lehrer des Kurses"
+        if self._data.attrib.get("UeLe") is not None and self._data.attrib.get("UeLe") != "":
+            return self._data.attrib["UeLe"]
+        else:
+            return None
+    
+    @property
+    def fach(self) -> str | None:
+        "Fach des Kurses"
+        if self._data.attrib.get("UeFa") is not None and self._data.attrib.get("UeFa") != "":
+            return self._data.attrib["UeFa"]
+        else:
+            return None
+    
+    @property
+    def gruppe(self) -> str | None:
+        "Gruppenbezeichnung des Kurses"
+        if self._data.attrib.get("UeGr") is not None and self._data.attrib.get("UeGr") != "":
+            return self._data.attrib["UeGr"]
+        else:
+            return None
+
+    @property
+    def kursnummer(self) -> int:
+        "Kursnummer des Kurses"
+        return int(self._data.text)